+++ stripe/src/main/java/com/stripe/android/PaymentController.kt
@@ -10,6 +10,8 @@ import com.stripe.android.model.ConfirmStripeIntentParams
 import com.stripe.android.model.PaymentIntent
 import com.stripe.android.model.Source
 import com.stripe.android.model.StripeIntent
+import com.stripe.android.model.WeChat
+import com.stripe.android.model.WeChatPayNextAction
 import com.stripe.android.networking.ApiRequest
 import com.stripe.android.view.AuthActivityStarter
 
@@ -51,6 +53,31 @@ internal interface PaymentController {
         authenticator: AlipayAuthenticator,
         requestOptions: ApiRequest.Options
     ): PaymentIntentResult
+
+    /**
+     * Confirm the Stripe Intent for WeChat Pay, return WeChat Pay params from intent's next action
+     *
+     * @param confirmPaymentIntentParams params to confirm the intent
+     * @param requestOptions options for [ApiRequest]
+     * @return the [WeChatPayNextAction] object encapsulating [PaymentIntent] and [WeChat]
+     *
+     * @throws AuthenticationException failure to properly authenticate yourself (check your key)
+     * @throws InvalidRequestException your request has invalid parameters
+     * @throws APIConnectionException failure to connect to Stripe's API
+     * @throws APIException any other type of problem (for instance, a temporary issue with Stripe's servers)
+     * @throws IllegalArgumentException if the payment intent's next action data is not for WeChat Pay
+     */
+    @Throws(
+        AuthenticationException::class,
+        InvalidRequestException::class,
+        APIConnectionException::class,
+        APIException::class,
+        IllegalArgumentException::class
+    )
+    suspend fun confirmWeChatPay(
+        confirmPaymentIntentParams: ConfirmPaymentIntentParams,
+        requestOptions: ApiRequest.Options
+    ): WeChatPayNextAction
 
     suspend fun startAuth(
         host: AuthActivityStarter.Host,