+++ stripe/src/main/java/com/stripe/android/StripePaymentController.kt
@@ -20,13 +20,15 @@ import com.stripe.android.model.Stripe3ds2AuthParams
 import com.stripe.android.model.Stripe3ds2AuthResult
 import com.stripe.android.model.Stripe3ds2Fingerprint
 import com.stripe.android.model.StripeIntent
+import com.stripe.android.model.WeChatPayNextAction
 import com.stripe.android.networking.AlipayRepository
 import com.stripe.android.networking.AnalyticsRequestExecutor
 import com.stripe.android.networking.AnalyticsRequestFactory
 import com.stripe.android.networking.ApiRequest
 import com.stripe.android.networking.DefaultAlipayRepository
 import com.stripe.android.networking.StripeRepository
-import com.stripe.android.payments.CustomTabsCapabilities
+import com.stripe.android.payments.BrowserCapabilities
+import com.stripe.android.payments.BrowserCapabilitiesSupplier
 import com.stripe.android.payments.DefaultPaymentFlowResultProcessor
 import com.stripe.android.payments.DefaultReturnUrl
 import com.stripe.android.payments.DefaultStripeChallengeStatusReceiver
@@ -99,8 +101,8 @@ internal class StripePaymentController internal constructor(
         } ?: PaymentRelayStarter.Legacy(host)
     }
 
-    private val isCustomTabsSupported: Boolean by lazy {
-        CustomTabsCapabilities(context).isSupported()
+    private val hasCompatibleBrowser: Boolean by lazy {
+        BrowserCapabilitiesSupplier(context).get() != BrowserCapabilities.Unknown
     }
 
     private val paymentBrowserAuthStarterFactory = { host: AuthActivityStarter.Host ->
@@ -108,7 +110,7 @@ internal class StripePaymentController internal constructor(
             PaymentBrowserAuthStarter.Modern(it)
         } ?: PaymentBrowserAuthStarter.Legacy(
             host,
-            isCustomTabsSupported,
+            hasCompatibleBrowser,
             defaultReturnUrl
         )
     }
@@ -191,6 +193,24 @@ internal class StripePaymentController internal constructor(
             authenticator,
             requestOptions
         )
+    }
+
+    override suspend fun confirmWeChatPay(
+        confirmPaymentIntentParams: ConfirmPaymentIntentParams,
+        requestOptions: ApiRequest.Options
+    ): WeChatPayNextAction {
+        confirmPaymentIntent(
+            confirmPaymentIntentParams,
+            requestOptions
+        ).let { paymentIntent ->
+            require(paymentIntent.nextActionData is StripeIntent.NextActionData.WeChatPayRedirect) {
+                "Unable to confirm Payment Intent with WeChatPay SDK"
+            }
+            return WeChatPayNextAction(
+                paymentIntent,
+                paymentIntent.nextActionData.weChat,
+            )
+        }
     }
 
     private suspend fun confirmPaymentIntent(
@@ -943,6 +963,9 @@ internal class StripePaymentController internal constructor(
                         analyticsRequestExecutor,
                         analyticsRequestFactory,
                         transaction,
+                        {
+                            transaction.close()
+                        },
                         workContext = workContext
                     ),
                     maxTimeout