+++ stripe/src/main/java/com/stripe/android/model/PaymentMethodCreateParams.kt
@@ -203,7 +203,8 @@ data class PaymentMethodCreateParams internal constructor(
         AfterpayClearpay("afterpay_clearpay"),
         Upi("upi"),
         Netbanking("netbanking"),
-        Blik("blik")
+        Blik("blik"),
+        WeChatPay("wechat_pay")
     }
 
     @Parcelize
@@ -765,6 +766,19 @@ data class PaymentMethodCreateParams internal constructor(
         ): PaymentMethodCreateParams {
             return PaymentMethodCreateParams(
                 type = Type.Blik,
+                billingDetails = billingDetails,
+                metadata = metadata
+            )
+        }
+
+        @JvmStatic
+        @JvmOverloads
+        fun createWeChatPay(
+            billingDetails: PaymentMethod.BillingDetails? = null,
+            metadata: Map<String, String>? = null
+        ): PaymentMethodCreateParams {
+            return PaymentMethodCreateParams(
+                type = Type.WeChatPay,
                 billingDetails = billingDetails,
                 metadata = metadata
             )