+++ stripe/src/main/java/com/stripe/android/payments/StripeBrowserLauncherActivity.kt
@@ -2,14 +2,11 @@ package com.stripe.android.payments
 
 import android.app.Activity
 import android.content.Intent
-import android.net.Uri
 import android.os.Bundle
 import androidx.activity.result.ActivityResult
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
-import androidx.browser.customtabs.CustomTabColorSchemeParams
-import androidx.browser.customtabs.CustomTabsIntent
 import com.stripe.android.auth.PaymentBrowserAuthContract
 import com.stripe.android.view.PaymentAuthWebViewActivity
 
@@ -19,16 +16,15 @@ import com.stripe.android.view.PaymentAuthWebViewActivity
  *
  * The eventual replacement for [PaymentAuthWebViewActivity].
  *
- * [StripeBrowserLauncherActivity] will only be used if Custom Tabs are enabled. See
- * [PaymentBrowserAuthContract.Args.shouldUseCustomTabs].
+ * [StripeBrowserLauncherActivity] will only be used when the following are true:
+ * - Custom Tabs are available or Chrome is installed
+ * - the confirmation request's `return_url` is set to [DefaultReturnUrl.value]
+ *
+ * See [BrowserCapabilities] and [PaymentBrowserAuthContract.Args.hasDefaultReturnUrl].
  */
 internal class StripeBrowserLauncherActivity : AppCompatActivity() {
     private val viewModel: StripeBrowserLauncherViewModel by viewModels {
         StripeBrowserLauncherViewModel.Factory(application)
-    }
-
-    private val customTabsCapabilities: CustomTabsCapabilities by lazy {
-        CustomTabsCapabilities(this)
     }
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -41,8 +37,6 @@ internal class StripeBrowserLauncherActivity : AppCompatActivity() {
             return
         }
 
-        val url = Uri.parse(args.url)
-
         setResult(
             Activity.RESULT_OK,
             viewModel.getResultIntent(args)
@@ -53,33 +47,9 @@ internal class StripeBrowserLauncherActivity : AppCompatActivity() {
             ::onResult
         )
 
-        val shouldUseCustomTabs = customTabsCapabilities.isSupported()
-        viewModel.logCapabilities(shouldUseCustomTabs)
-
-        if (shouldUseCustomTabs) {
-            val customTabColorSchemeParams = args.statusBarColor?.let { statusBarColor ->
-                CustomTabColorSchemeParams.Builder()
-                    .setToolbarColor(statusBarColor)
-                    .build()
-            }
-
-            // use Custom Tabs
-            val customTabsIntent = CustomTabsIntent.Builder()
-                .setShareState(CustomTabsIntent.SHARE_STATE_OFF)
-                .also {
-                    if (customTabColorSchemeParams != null) {
-                        it.setDefaultColorSchemeParams(customTabColorSchemeParams)
-                    }
-                }
-                .build()
-            customTabsIntent.intent.data = url
-            launcher.launch(customTabsIntent.intent)
-        } else {
-            // use default device browser
-            launcher.launch(
-                Intent(Intent.ACTION_VIEW, url)
-            )
-        }
+        launcher.launch(
+            viewModel.createLaunchIntent(args)
+        )
     }
 
     private fun onResult(activityResult: ActivityResult) {