+++ stripe/src/test/java/com/stripe/android/PaymentMethodEndToEndTest.kt
@@ -324,4 +324,18 @@ internal class PaymentMethodEndToEndTest {
         assertThat(paymentMethod?.type)
             .isEqualTo(PaymentMethod.Type.Blik)
     }
+
+    @Test
+    fun createPaymentMethod_withWeChatPay_shouldCreateObject() {
+        val params = PaymentMethodCreateParams.createWeChatPay()
+        val paymentMethod =
+            Stripe(
+                context,
+                ApiKeyFixtures.WECHAT_PAY_PUBLISHABLE_KEY,
+                betas = setOf(StripeApiBeta.WeChatPayV1)
+            )
+                .createPaymentMethodSynchronous(params)
+        assertThat(paymentMethod?.type)
+            .isEqualTo(PaymentMethod.Type.WeChatPay)
+    }
 }